requirements-dev.txt
-r requirements.txt
black==26.3.1
coverage==7.13.5
django-debug-toolbar==6.3.0
djlint==1.36.4
flake8==7.3.0
isort==8.0.1
pip-audit==2.10.0
pytest-django==4.12.0
pytest==9.0.3
pytest-cov==7.1.0
